+++ js/packages/openinference-instrumentation-beeai/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @arizeai/openinference-instrumentation-beeai
 
+## 1.5.21
+
+### Patch Changes
+
+- Updated dependencies [237ce2b]
+  - @arizeai/openinference-semantic-conventions@2.7.0
+  - @arizeai/openinference-core@2.5.1
+
 ## 1.5.20
 
 ### Patch Changes