+++ js/packages/openinference-instrumentation-mcp/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @arizeai/openinference-instrumentation-mcp
 
+## 0.2.23
+
+### Patch Changes
+
+- Updated dependencies [237ce2b]
+  - @arizeai/openinference-semantic-conventions@2.7.0
+  - @arizeai/openinference-core@2.5.1
+
 ## 0.2.22
 
 ### Patch Changes